Modified sessions to have an expiration date
This was necessary to make the sessions feature useful. Basically,
at the moment the situation is that a database administrator can
give anyone any permissions via an active session, and that session
does not have to be the one that the user is currently using in order
to activate that role.

Another change which is present is the modification of updateUserRole
to not allow you to change the role of yourself, as that would lead to
an exception in the current implementation and I did not have time to
fix it.

Moved the authentication code into a new file